Upstream version 2.3~dfsg

+++ pjlib/src/pj/errno.c
@@ -1,4 +1,4 @@
-/* $Id: errno.c 4613 2013-10-08 09:08:13Z bennylp $ */
+/* $Id: errno.c 4761 2014-02-24 09:02:44Z nanang $ */
 /* 
  * Copyright (C) 2008-2011 Teluu Inc. (http://www.teluu.com)
  * Copyright (C) 2003-2008 Benny Prijono <benny@prijono.org>
@@ -109,7 +109,7 @@ static int pjlib_error(pj_status_t code, char *buf, pj_size_t size)
 
     len = pj_ansi_snprintf( buf, size, "Unknown pjlib error %d", code);
     if (len < 1 || len >= (int)size)
-	len = size - 1;
+	len = (int)(size - 1);
     return len;
 }
 
@@ -205,7 +205,7 @@ PJ_DEF(pj_str_t) pj_strerror( pj_status_t statcode,
     }
 
     if (len < 1 || len >= (int)bufsize) {
-	len = bufsize - 1;
+	len = (int)(bufsize - 1);
 	buf[len] = '\0';
     }
